+++ src/main/java/com/example/homework50/dao/SubscriptionsDao.java
@@ -15,14 +15,14 @@ import java.time.LocalDateTime;
 public class SubscriptionsDao {
     private final JdbcTemplate jdbcTemplate;
 
-    public void createSubscriptions(int whoSubscribes, int whoIsSubscribedTo, LocalDateTime subDateTime) {
+    public void createSubscriptions(Long whoSubscribes, Long whoIsSubscribedTo, LocalDateTime subDateTime) {
         String sql = "insert into subscriptions (who_subscribes_id, who_is_subscribed_to_id, date_time) values(?,?,?);";
         KeyHolder keyHolder = new GeneratedKeyHolder();
 
         jdbcTemplate.update(con -> {
             PreparedStatement ps = con.prepareStatement(sql, new String[]{"id"});
-            ps.setInt(1, whoSubscribes);
-            ps.setInt(2, whoIsSubscribedTo);
+            ps.setLong(1, whoSubscribes);
+            ps.setLong(2, whoIsSubscribedTo);
             ps.setObject(3, subDateTime);
             return ps;
         }, keyHolder);
